+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-## vNext
+## 0.6.0
 
 ### Features